integration tests: move bridge integration tests to a separate file
